✨ feat(toast): refactor ToastProvider integration and update exports
